requirements.txt
wheel==0.32.3
astroid==2.0.4
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
flask-marshmallow==0.9.0
flask-paginate==0.5.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.17.0
marshmallow-sqlalchemy==0.15.0
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.7.6.1
pycparser==2.19
pylint==2.1.1
six==1.11.0
SQLAlchemy==1.2.13
typed-ast==1.1.0
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
